move path decoder to another file
